Move section-specific analytics modules (signup, cart, transactions) to lib/analytics
