+++ client/lib/analytics/signup.js
