+++ client/signup/main.jsx
@@ -33,7 +33,7 @@ import {
 	recordSignupComplete,
 	recordSignupStep,
 	recordSignupInvalidStep,
-} from 'lib/signup/analytics';
+} from 'lib/analytics/signup';
 import DocumentHead from 'components/data/document-head';
 import LocaleSuggestions from 'components/locale-suggestions';
 import SignupProcessingScreen from 'signup/processing-screen';
